Allow to polish plugins before loading
This change allows users to manipulate the lua table passed to packer to
load plugins, allowing to override every aspect of the configuration.

Note that this is very low level and not suited to change plugin
configuration since there the configuration suggested by AstroVim is
lost at this point.

Fixes #113, Fixes #114